Actualización Repository y entities

+++ src/main/kotlin/ProyectoCitas/Domains/requests/AppointmentReq.kt
@@ -5,7 +5,7 @@ data class AppointmentReq(
     @field:NotNull("El horario de la cita es requerido")
     val horario: String,
 
-    val id_doctor: Int,
+    val id_doctor: Long,
     val identif_paciente: String
     
 ){
